requirements.txt
langchain==0.3.27
langchain-community==0.3.31
langchain-huggingface==0.1.0
faiss-cpu==1.8.0
sentence-transformers==3.1.1
mlflow==2.17.0
python-dotenv==1.0.1
transformers==4.46.3
torch==2.5.1
accelerate==1.1.1
